requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
yt-dlp
scenedetect[opencv]
ultralytics
openai-whisper
sentence-transformers
groq
python-dotenv
ffmpeg-python
numpy
Pillow
